requirements.txt
pyyaml
easydict
transformers
h5py
tqdm
numpy
nltk